+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 0.1.1
 
 - Fixed exports
+- Created test annotation in example
 
 ## 0.1.0